refactor: #149
command 패키지 domain으로 변경
